get_data_for_analytics_bot/main.py
import os
import asyncio
import logging

from aiogram import Bot, Dispatcher
from dotenv import load_dotenv
from app.handlers import router


load_dotenv()
bot = Bot(token=os.getenv('TOKEN'))
dp = Dispatcher()


async def main():
    dp.include_router(router)
    await dp.start_polling(bot)


if __name__ == '__main__':
    logging.basicConfig(level=logging.INFO)
    try:
        asyncio.run(main())
    except KeyboardInterrupt:
        print('Exit')
